backend/textbooks/management/commands/fetch_textbooks.py
import requests
from bs4 import BeautifulSoup
from django.core.management.base import BaseCommand
from textbooks.models import Textbook


class Command(BaseCommand):
    help = 'Fetch and seed textbook data from CBSE (NCERT), ICSE, and Karnataka State Board'

    def handle(self, *args, **kwargs):
        self.fetch_ncert_cbse()
        self.fetch_karnataka()
        self.fetch_icse()

    def fetch_ncert_cbse(self):
        """
        NCERT provides a PHP-based textbook portal.
        Target URL pattern: https://ncert.nic.in/textbook.php?{subjectCode}{grade}-{chapterRange}
        Scrape the textbook listing page and extract PDF links per grade and subject.
        Store board='CBSE' for all NCERT books.
        """
        base_url = 'https://ncert.nic.in/textbook.php'
        _ = (requests, BeautifulSoup, Textbook, base_url)
        # TODO: Implement scraping logic based on live NCERT markup.
        self.stdout.write('Fetched CBSE (NCERT) textbooks.')

    def fetch_karnataka(self):
        """
        Karnataka Textbook Society (KTBS) public portal: https://ktbs.kar.nic.in
        Scrape the 'Free Textbooks' section for PDF download links per class and subject.
        Store board='KA_STATE' for all Karnataka books.
        Use requests + BeautifulSoup. Handle relative URLs carefully.
        """
        base_url = 'https://ktbs.kar.nic.in'
        _ = (requests, BeautifulSoup, Textbook, base_url)
        # TODO: Implement scraping logic based on live KTBS markup.
        self.stdout.write('Fetched Karnataka State Board textbooks.')

    def fetch_icse(self):
        """
        CISCE publishes study material at: https://cisce.org/publications.aspx
        Scrape the page for ICSE/ISC book links.
        Store board='ICSE' for all entries.
        """
        base_url = 'https://cisce.org/publications.aspx'
        _ = (requests, BeautifulSoup, Textbook, base_url)
        # TODO: Implement scraping logic based on live CISCE markup.
        self.stdout.write('Fetched ICSE textbooks.')
